Converting to new FW format
