+++ apis/mai400-api/src/messages/mod.rs
@@ -17,7 +17,41 @@
 mod rx;
 mod tx;
 
+use byteorder::{LittleEndian, WriteBytesExt};
 pub use self::rx::*;
 pub use self::tx::*;
 
 pub const SYNC: u16 = 0xEB90;
+pub const HDR_SZ: usize = 6;
+pub const FRAME_SZ: usize = HDR_SZ + 2;
+
+#[repr(C, packed)]
+pub struct MessageHeader {
+    pub sync: u16,
+    pub data_len: u16,
+    pub msg_id: u8,
+    pub addr: u8,
+}
+
+impl Default for MessageHeader {
+    fn default() -> Self {
+        MessageHeader {
+            sync: SYNC,
+            data_len: 0,
+            msg_id: 0,
+            addr: 0,
+        }
+    }
+}
+
+impl MessageHeader {
+    fn serialize(&self) -> Vec<u8> {
+        let mut vec = vec![];
+
+        vec.write_u16::<LittleEndian>(self.sync).unwrap();
+        vec.write_u16::<LittleEndian>(self.data_len).unwrap();
+        vec.push(self.msg_id);
+        vec.push(self.addr);
+        vec
+    }
+}
