+++ apis/mai400-api/src/messages/rx.rs
@@ -15,9 +15,10 @@
 //
 
 use nom::*;
+use super::*;
 
 pub struct StandardTelemetry {
-    pub sync: u16,
+    pub hdr: MessageHeader,
     pub tlm_counter: u8,
     pub gps_time: u32,
     pub time_subsec: u8,
@@ -60,31 +61,31 @@ pub struct StandardTelemetry {
     pub gc_torque_coil_cmd_0: char,
     pub gc_torque_coil_cmd_1: char,
     pub gc_torque_coil_cmd_2: char,
-    pub qbo_cmd_0: i16,
-    pub qbo_cmd_1: i16,
-    pub qbo_cmd_2: i16,
-    pub qbo_cmd_3: i16,
-    pub qbo_hat_0: i16,
-    pub qbo_hat_1: i16,
-    pub qbo_hat_2: i16,
-    pub qbo_hat_3: i16,
+    pub qbo_cmd_0: i32,
+    pub qbo_cmd_1: i32,
+    pub qbo_cmd_2: i32,
+    pub qbo_cmd_3: i32,
+    pub qbo_hat_0: i32,
+    pub qbo_hat_1: i32,
+    pub qbo_hat_2: i32,
+    pub qbo_hat_3: i32,
     pub angle_to_go: f32,
-    pub q_error_0: i16,
-    pub q_error_1: i16,
-    pub q_error_2: i16,
-    pub q_error_3: i16,
+    pub q_error_0: i32,
+    pub q_error_1: i32,
+    pub q_error_2: i32,
+    pub q_error_3: i32,
     pub omega_b_0: f32,
     pub omega_b_1: f32,
     pub omega_b_2: f32,
     pub rotating_variable_a: u32,
     pub rotating_variable_b: u32,
     pub rotating_variable_c: u32,
-    pub nb_0: i16,
-    pub nb_1: i16,
-    pub nb_2: i16,
-    pub neci_0: i16,
-    pub neci_1: i16,
-    pub neci_2: i16,
+    pub nb_0: i32,
+    pub nb_1: i32,
+    pub nb_2: i32,
+    pub neci_0: i32,
+    pub neci_1: i32,
+    pub neci_2: i32,
     pub crc: u16,
 }
 
@@ -97,6 +98,9 @@ impl StandardTelemetry {
 named!(standardtelem(&[u8]) -> StandardTelemetry,
     do_parse!(
         sync: le_u16 >>
+        data_len: le_u16 >>
+        msg_id: le_u8 >>
+        addr: le_u8 >>
         tlm_counter: le_u8 >>
         gps_time: le_u32 >>
         time_subsec: le_u8 >>
@@ -139,34 +143,39 @@ named!(standardtelem(&[u8]) -> StandardTelemetry,
         gc_torque_coil_cmd_0: le_u8 >>
         gc_torque_coil_cmd_1: le_u8 >>
         gc_torque_coil_cmd_2: le_u8 >>
-        qbo_cmd_0: le_i16 >>
-        qbo_cmd_1: le_i16 >>
-        qbo_cmd_2: le_i16 >>
-        qbo_cmd_3: le_i16 >>
-        qbo_hat_0: le_i16 >>
-        qbo_hat_1: le_i16 >>
-        qbo_hat_2: le_i16 >>
-        qbo_hat_3: le_i16 >>
+        qbo_cmd_0: le_i32 >>
+        qbo_cmd_1: le_i32 >>
+        qbo_cmd_2: le_i32 >>
+        qbo_cmd_3: le_i32 >>
+        qbo_hat_0: le_i32 >>
+        qbo_hat_1: le_i32 >>
+        qbo_hat_2: le_i32 >>
+        qbo_hat_3: le_i32 >>
         angle_to_go: le_f32 >>
-        q_error_0: le_i16 >>
-        q_error_1: le_i16 >>
-        q_error_2: le_i16 >>
-        q_error_3: le_i16 >>
+        q_error_0: le_i32 >>
+        q_error_1: le_i32 >>
+        q_error_2: le_i32 >>
+        q_error_3: le_i32 >>
         omega_b_0: le_f32 >>
         omega_b_1: le_f32 >>
         omega_b_2: le_f32 >>
         rotating_variable_a: le_u32 >>
         rotating_variable_b: le_u32 >>
         rotating_variable_c: le_u32 >>
-        nb_0: le_i16 >>
-        nb_1: le_i16 >>
-        nb_2: le_i16 >>
-        neci_0: le_i16 >>
-        neci_1: le_i16 >>
-        neci_2: le_i16 >>
+        nb_0: le_i32 >>
+        nb_1: le_i32 >>
+        nb_2: le_i32 >>
+        neci_0: le_i32 >>
+        neci_1: le_i32 >>
+        neci_2: le_i32 >>
         crc: le_u16 >>
         (StandardTelemetry{
-            sync,
+            hdr: MessageHeader {
+                sync, 
+                data_len, 
+                msg_id, 
+                addr
+            },
             tlm_counter,
             gps_time,
             time_subsec,
@@ -240,9 +249,7 @@ named!(standardtelem(&[u8]) -> StandardTelemetry,
 );
 
 pub struct RawIMU {
-    pub sync: u16,
-    pub data_len: u16,
-    pub packet_type: u16,
+    pub hdr: MessageHeader,
     pub accel: [i16; 3],
     pub gyro: [i16; 3],
     pub gyro_temp: u8,
@@ -259,7 +266,8 @@ named!(raw_imu(&[u8]) -> RawIMU,
     do_parse!(
         sync: le_u16 >>
         data_len: le_u16 >>
-        packet_type: le_u16 >>
+        msg_id: le_u8 >>
+        addr: le_u8 >>
         accel_x: le_i16 >>
         accel_y: le_i16 >>
         accel_z: le_i16 >>
@@ -269,9 +277,12 @@ named!(raw_imu(&[u8]) -> RawIMU,
         gyro_temp: le_u8 >>
         crc: le_u16 >>
         (RawIMU {
-                sync,
-                data_len,
-                packet_type,
+                hdr: MessageHeader {
+                    sync, 
+                    data_len, 
+                    msg_id, 
+                    addr
+                },
                 accel: [accel_x ,accel_y, accel_z],
                 gyro: [gyro_x, gyro_y, gyro_z],
                 gyro_temp,
@@ -281,9 +292,7 @@ named!(raw_imu(&[u8]) -> RawIMU,
 );
 
 pub struct IREHSTelemetry {
-    pub sync: u16,
-    pub data_len: u16,
-    pub packet_type: u16,
+    pub hdr: MessageHeader,
     pub thermopiles_a: [u16; 4],
     pub thermopiles_b: [u16; 4],
     pub temp_a: [u16; 4],
@@ -304,7 +313,8 @@ named!(irehs_telem(&[u8]) -> IREHSTelemetry,
     do_parse!(
         sync: le_u16 >>
         data_len: le_u16 >>
-        packet_type: le_u16 >>
+        msg_id: le_u8 >>
+        addr: le_u8 >>
         thermopiles_a_earth_limb: le_u16 >>
         thermopiles_a_earth_ref: le_u16 >>
         thermopiles_a_space_ref: le_u16 >>
@@ -333,9 +343,12 @@ named!(irehs_telem(&[u8]) -> IREHSTelemetry,
         solution_degraded_wide_fov_b: le_u8 >>
         crc: le_u16 >>
         (IREHSTelemetry {
-                sync,
-                data_len,
-                packet_type,
+                hdr: MessageHeader {
+                    sync, 
+                    data_len, 
+                    msg_id, 
+                    addr
+                },
                 thermopiles_a: [
                     thermopiles_a_earth_limb,
                     thermopiles_a_earth_ref,
@@ -377,8 +390,83 @@ named!(irehs_telem(&[u8]) -> IREHSTelemetry,
     )
 );
 
+
+pub struct ConfigInfo {
+    pub hdr: MessageHeader,
+    pub model: u16,
+    pub serial: u16,
+    pub major: u8,
+    pub minor: u8,
+    pub build: u16,
+    pub n_ehs: u8,
+    pub ehs_type: EHSType,
+    pub n_st: u8,
+    pub st_type: StarTracker,
+    pub crc: u16,
+}
+
+impl ConfigInfo {
+    pub fn new(msg: &[u8]) -> Self {
+        configinfo(msg).unwrap().1
+    }
+}
+
+named!(configinfo(&[u8]) -> ConfigInfo,
+    do_parse!(
+        sync: le_u16 >>
+        data_len: le_u16 >>
+        msg_id: le_u8 >>
+        addr: le_u8 >>
+        model: le_u16 >>
+        serial: le_u16 >>
+        major: le_u8 >>
+        minor: le_u8 >>
+        build: le_u16 >>
+        n_ehs: le_u8 >> 
+        ehs_type: le_u8 >>
+        n_st: le_u8 >>
+        st_type: le_u8 >>
+        crc: le_u16 >>
+        (ConfigInfo{
+                hdr: MessageHeader {
+                    sync, 
+                    data_len, 
+                    msg_id, 
+                    addr
+                }, 
+                model, 
+                serial, 
+                major, 
+                minor,
+                build, 
+                n_ehs,
+                ehs_type: match ehs_type {
+                    0 => EHSType::Internal,
+                    _ => EHSType::External
+                },
+                n_st,
+                st_type: match st_type {
+                    0 => StarTracker::MAISextant,
+                    _ => StarTracker::Vectronic,
+                }, 
+                crc})
+    )
+);
+
+
 pub enum Response {
     StdTelem(StandardTelemetry),
     IMU(RawIMU),
     IREHS(IREHSTelemetry),
+    Config(ConfigInfo),
+}
+
+pub enum EHSType {
+    Internal,
+    External,
+}
+
+pub enum StarTracker {
+    MAISextant,
+    Vectronic,
 }
