+++ apis/mai400-api/src/messages/tx.rs
@@ -21,8 +21,29 @@ pub trait Message {
     fn serialize(&self) -> Vec<u8>;
 }
 
+pub struct GetInfo([u8; 6]);
+
+impl Default for GetInfo {
+    fn default() -> Self {
+        let mut array = [0; 6];
+        array[0] = 0x90; // SYNC byte 1
+        array[1] = 0xEB; // SYNC byte 2
+        array[2] = 0x0; // Data_len byte 1
+        array[3] = 0x0; // Data_len byte 2
+        array[4] = 0x1D; // Msg_id
+        array[5] = 0x0; // Addr
+        GetInfo(array)
+    }
+}
+
+impl Message for GetInfo {
+    fn serialize(&self) -> Vec<u8> {
+        self.0.to_vec()
+    }
+}
+
 pub struct SetAcsMode {
-    //pub hdr: MessageHeader,
+    pub hdr: MessageHeader,
     pub mode: u8,
     pub sec_vec: i32,
     pub pri_axis: i32,
@@ -33,6 +54,10 @@ pub struct SetAcsMode {
 impl Default for SetAcsMode {
     fn default() -> Self {
         SetAcsMode {
+            hdr: MessageHeader {
+                data_len: 17,
+                ..Default::default()
+            },
             mode: 0,
             sec_vec: 0,
             pri_axis: 0,
@@ -46,7 +71,7 @@ impl Message for SetAcsMode {
     fn serialize(&self) -> Vec<u8> {
         let mut vec = vec![];
 
-        //vec.append(&mut self.hdr.serialize());
+        vec.append(&mut self.hdr.serialize());
         vec.push(self.mode);
         vec.write_i32::<LittleEndian>(self.sec_vec).unwrap();
         vec.write_i32::<LittleEndian>(self.pri_axis).unwrap();
@@ -57,17 +82,18 @@ impl Message for SetAcsMode {
 }
 
 pub struct SetGPSTime {
-    //pub hdr: MessageHeader,
-    pub sync: u16,
-    pub id: u8,
+    pub hdr: MessageHeader,
     pub gps_time: u32,
 }
 
 impl Default for SetGPSTime {
     fn default() -> Self {
         SetGPSTime {
-            sync: SYNC,
-            id: 0x44,
+            hdr: MessageHeader {
+                data_len: 4,
+                msg_id: 0x44,
+                ..Default::default()
+            },
             gps_time: 0,
         }
     }
@@ -77,19 +103,14 @@ impl Message for SetGPSTime {
     fn serialize(&self) -> Vec<u8> {
         let mut vec = vec![];
 
-        //vec.append(&mut self.hdr.serialize());
-        vec.write_u16::<LittleEndian>(self.sync).unwrap();
-        vec.push(self.id);
+        vec.append(&mut self.hdr.serialize());
         vec.write_u32::<LittleEndian>(self.gps_time).unwrap();
-        vec.append(&mut vec![0; 31]); //TODO: Remove for newer FW version
         vec
     }
 }
 
 pub struct SetRV {
-    //pub hdr: MessageHeader,
-    pub sync: u16,
-    pub id: u8,
+    pub hdr: MessageHeader,
     pub eci_pos_x: f32,
     pub eci_pos_y: f32,
     pub eci_pos_z: f32,
@@ -102,8 +123,11 @@ pub struct SetRV {
 impl Default for SetRV {
     fn default() -> Self {
         SetRV {
-            sync: SYNC,
-            id: 0x41,
+            hdr: MessageHeader {
+                data_len: 28,
+                msg_id: 0x41,
+                ..Default::default()
+            },
             eci_pos_x: 0.0,
             eci_pos_y: 0.0,
             eci_pos_z: 0.0,
@@ -119,9 +143,7 @@ impl Message for SetRV {
     fn serialize(&self) -> Vec<u8> {
         let mut vec = vec![];
 
-        //vec.append(&mut self.hdr.serialize());
-        vec.write_u16::<LittleEndian>(self.sync).unwrap();
-        vec.push(self.id);
+        vec.append(&mut self.hdr.serialize());
         vec.write_f32::<LittleEndian>(self.eci_pos_x).unwrap();
         vec.write_f32::<LittleEndian>(self.eci_pos_y).unwrap();
         vec.write_f32::<LittleEndian>(self.eci_pos_z).unwrap();
@@ -129,33 +151,48 @@ impl Message for SetRV {
         vec.write_f32::<LittleEndian>(self.eci_vel_y).unwrap();
         vec.write_f32::<LittleEndian>(self.eci_vel_z).unwrap();
         vec.write_u32::<LittleEndian>(self.time_epoch).unwrap();
-        vec.append(&mut vec![0; 6]); //TODO: Remove for newer FW version
         vec
     }
 }
 
-pub struct RequestReset([u8; 40]);
+pub struct RequestReset(pub [u8; 6]);
 
 impl Default for RequestReset {
     fn default() -> Self {
-        let mut array = [0; 40];
-        array[0] = 0x90;
-        array[1] = 0xEB;
-        array[2] = 0x5A;
-        //TODO: Checksum
+        let mut array = [0; 6];
+        array[0] = 0x90; // SYNC byte 1
+        array[1] = 0xEB; // SYNC byte 2
+        array[2] = 0x0; // Data_len byte 1
+        array[3] = 0x0; // Data_len byte 2
+        array[4] = 0x5A; // Msg_id
+        array[5] = 0x0; // Addr
         RequestReset(array)
     }
 }
 
-pub struct ConfirmReset([u8; 40]);
+impl Message for RequestReset {
+    fn serialize(&self) -> Vec<u8> {
+        self.0.to_vec()
+    }
+}
+
+pub struct ConfirmReset([u8; 6]);
 
 impl Default for ConfirmReset {
     fn default() -> Self {
-        let mut array = [0; 40];
-        array[0] = 0x90;
-        array[1] = 0xEB;
-        array[2] = 0xF1;
-        //TODO: Checksum
+        let mut array = [0; 6];
+        array[0] = 0x90; // SYNC byte 1
+        array[1] = 0xEB; // SYNC byte 2
+        array[2] = 0x0; // Data_len byte 1
+        array[3] = 0x0; // Data_len byte 2
+        array[4] = 0xF1; // Msg_id
+        array[5] = 0x0; // Addr
         ConfirmReset(array)
+    }
+}
+
+impl Message for ConfirmReset {
+    fn serialize(&self) -> Vec<u8> {
+        self.0.to_vec()
     }
 }
